cart without login

+++ src/components/addCourse/courseChapters.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import FileUploader from 'react-files';
-import { Form, Input, Button, Table, Icon, Accordion, Header, Confirm } from 'semantic-ui-react'
+import { Form, Input, Button, Table, Icon, Accordion, Header, Confirm, Label } from 'semantic-ui-react'
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { withRouter } from 'react-router-dom'
@@ -18,7 +18,8 @@ class CourseChapters extends Component {
         chapterId: 0,
         currentFiles: [],
         deleteFileId: [],
-        showConfirm: false
+        showConfirm: false,
+        submitted: false
     }
     componentWillMount() {
         this.props.getChapters(Number(this.props.match.params.id))
@@ -30,25 +31,29 @@ class CourseChapters extends Component {
         this.setState({ chapterFiles: files })
     }
     addChapter = () => {
-        let data = new FormData();
-        data.append('chapter_title', this.state.chapterTitle);
-        if (this.state.chapterFiles.length > 0) {
-            this.state.chapterFiles.forEach(file => {
-                data.append('chapter_files', file);
-            })
-        }
-        data.append('CourseId', Number(this.props.match.params.id))
-        this.setState(oldState => ({
-            chapterTitle: '',
-            chapterFiles: [],
-            key: oldState.key + 1
-        }))
-        const config = {
-            headers: {
-                'content-type': 'multipart/form-data'
+        this.setState({ submitted: true });
+        if (this.state.chapterTitle !== '') {
+            let data = new FormData();
+            data.append('chapter_title', this.state.chapterTitle);
+            if (this.state.chapterFiles.length > 0) {
+                this.state.chapterFiles.forEach(file => {
+                    data.append('chapter_files', file);
+                })
             }
+            data.append('CourseId', Number(this.props.match.params.id))
+            this.setState(oldState => ({
+                chapterTitle: '',
+                chapterFiles: [],
+                key: oldState.key + 1,
+                submitted: false
+            }))
+            const config = {
+                headers: {
+                    'content-type': 'multipart/form-data'
+                }
+            }
+            this.props.addChapter(data, config)
         }
-        this.props.addChapter(data, config)
     }
     editChapter = (chapter) => {
         this.setState({ chapterTitle: chapter.chapter_title, currentFiles: [...chapter.ChapterFiles], editing: true, chapterId: chapter.id })
@@ -58,34 +63,39 @@ class CourseChapters extends Component {
         this.toggleConfirm()
     }
     updateChapter = () => {
+        this.setState({ submitted: true });
         let { deleteFileId, chapterId, chapterFiles, chapterTitle } = this.state;
-        if (deleteFileId.length > 0) {
-            deleteFileId.forEach(id => {
-                this.props.deleteFile(id, chapterId)
-            })
-        }
-        let data = new FormData();
-        data.append('chapter_title', chapterTitle);
-        if (chapterFiles.length > 0) {
-            this.state.chapterFiles.forEach(file => {
-                data.append('chapter_files', file);
-            })
-        }
-        data.append('CourseId', Number(this.props.match.params.id))
-        this.setState(oldState => ({
-            chapterTitle: '',
-            chapterFiles: [],
-            key: oldState.key + 1,
-            currentFiles: [],
-            deleteFileId: [],
-            editing: false
-        }))
-        const config = {
-            headers: {
-                'content-type': 'multipart/form-data'
+        if (chapterTitle !== '') {
+            if (deleteFileId.length > 0) {
+                deleteFileId.forEach(id => {
+                    this.props.deleteFile(id, chapterId)
+                })
             }
+
+            let data = new FormData();
+            data.append('chapter_title', chapterTitle);
+            if (chapterFiles.length > 0) {
+                this.state.chapterFiles.forEach(file => {
+                    data.append('chapter_files', file);
+                })
+            }
+            data.append('CourseId', Number(this.props.match.params.id))
+            this.setState(oldState => ({
+                chapterTitle: '',
+                chapterFiles: [],
+                key: oldState.key + 1,
+                currentFiles: [],
+                deleteFileId: [],
+                editing: false,
+                submitted: false
+            }))
+            const config = {
+                headers: {
+                    'content-type': 'multipart/form-data'
+                }
+            }
+            this.props.editChapter(chapterId, { chapterTitle }, data, config)
         }
-        this.props.editChapter(chapterId, { chapterTitle }, data, config)
     }
     deleteFile = (id) => {
         this.setState(oldState => ({ deleteFileId: [...oldState.deleteFileId, id] }))
@@ -93,7 +103,6 @@ class CourseChapters extends Component {
         let i = currentFiles.findIndex(file => file.id === id)
         currentFiles.splice(i, 1);
         this.setState({ currentFiles });
-        // this.props.deleteFile(id, chapterId)
     }
     removeFile = (id) => {
         let { chapterFiles } = this.state;
@@ -107,10 +116,11 @@ class CourseChapters extends Component {
                 <div style={{ boxShadow: '2px 3px 2px 2px lightgrey', padding: '15px', marginBottom: '15px' }}>
                     <Form>
                         <Form.Field>
-                            <Input type='text' placeholder='Chapter title' value={this.state.chapterTitle} onChange={(e) => this.setState({ chapterTitle: e.target.value })} />
+                            <Input type='text' placeholder='Chapter title' value={this.state.chapterTitle} onChange={(e) => this.setState({ chapterTitle: e.target.value, submitted: false })} />
+                            {this.state.submitted && this.state.chapterTitle === '' ? <Label basic color='red' pointing>This field cannot be empty</Label> : null}
                         </Form.Field>
                         <Form.Field style={{ textAlign: 'center' }}>
-                            <FileUploader key={this.state.key} clickable multiple onChange={this.onFilesChange}>
+                            <FileUploader key={this.state.key} clickable multiple onChange={this.onFilesChange} accepts={['video/mp4', 'video/ogg', '.pdf']}>
                                 <Button style={{ borderRadius: '0px' }} color='linkedin' content='Click to upload files' />
                             </FileUploader>
                             {this.state.chapterFiles.length > 0 || this.state.currentFiles.length > 0 ? <Table celled size='small' >
@@ -118,7 +128,7 @@ class CourseChapters extends Component {
                                     {this.state.chapterFiles.map((file, i) => {
                                         return <Table.Row textAlign='center' key={i}>
                                             <Table.Cell>
-                                                {file.type.match('video') ? <Icon size='big' name='video' /> : <Icon size='big' name='file' />}
+                                                {file.type.match('video') ? <Icon size='big' name='video play' /> : <Icon size='big' name='file' />}
                                             </Table.Cell>
                                             <Table.Cell>{file.name}</Table.Cell>
                                             <Table.Cell>
@@ -130,7 +140,7 @@ class CourseChapters extends Component {
                                         this.state.currentFiles.map((file, i) => {
                                             return <Table.Row textAlign='center' key={i}>
                                                 <Table.Cell>
-                                                    {file.file_type.match('video') ? <Icon size='big' name='video' /> : <Icon size='big' name='file' />}
+                                                    {file.file_type.match('video') ? <Icon size='big' name='video play' /> : <Icon size='big' name='file' />}
                                                 </Table.Cell>
                                                 <Table.Cell>{file.file_name}</Table.Cell>
                                                 <Table.Cell>
@@ -184,7 +194,7 @@ class CourseChapters extends Component {
                                         <Table.Body>
                                             {chapter.ChapterFiles.map((file, i) => {
                                                 return <Table.Row key={i}>
-                                                    <Table.Cell textAlign='left'>{file.file_name}</Table.Cell>
+                                                    <Table.Cell> {file.file_type.match('video') ? <Icon name='video play' /> : <Icon name='file' />} {file.file_name}</Table.Cell>
                                                 </Table.Row>
                                             })}
                                         </Table.Body>
